Restore SigScanner as GenericSigScanner
- Expose `GenericSigScanner` for plugins as public API for plugins that need to scan arbitrary process modules.
- Convert SigScanner into a game-specific wrapper for `GenericSigScanner`.
- `ISigScanner` will resolve the game-specific `SigScanner`
- Convert non-game sig systems to the generic scanner.
